Avoid some EB stuff on non-EB runs
