+++ Source/InitEB.cpp
@@ -1,5 +1,6 @@
 #include <memory>
 
+#include "hydro_redistribution.H"
 #include "EB.H"
 #include "prob.H"
 #include "Utilities.H"
@@ -20,6 +21,10 @@ PeleC::ebInitialized()
 void
 PeleC::init_eb()
 {
+  if (!eb_in_domain) {
+    return;
+  }
+
   // Build the geometry information; this is done for each new set of grids
   initialize_eb2_structs();
 
@@ -42,19 +47,13 @@ PeleC::initialize_eb2_structs()
     std::is_standard_layout<EBBndryGeom>::value,
     "EBBndryGeom is not standard layout");
 
-  const amrex::MultiCutFab* bndrycent;
-  std::array<const amrex::MultiCutFab*, AMREX_SPACEDIM> eb2areafrac;
-  std::array<const amrex::MultiCutFab*, AMREX_SPACEDIM> facecent;
-
   const auto& ebfactory =
     dynamic_cast<amrex::EBFArrayBoxFactory const&>(Factory());
 
   // These are the data sources
-  vfrac.clear();
-  vfrac.define(grids, dmap, 1, numGrow(), amrex::MFInfo(), Factory());
   amrex::MultiFab::Copy(vfrac, ebfactory.getVolFrac(), 0, 0, 1, numGrow());
-  bndrycent = &(ebfactory.getBndryCent());
-  eb2areafrac = ebfactory.getAreaFrac();
+  const amrex::MultiCutFab* bndrycent = &(ebfactory.getBndryCent());
+  areafrac = ebfactory.getAreaFrac();
   facecent = ebfactory.getFaceCent();
 
   // First pass over fabs to fill sparse per cut-cell ebg structures
@@ -121,13 +120,12 @@ PeleC::initialize_eb2_structs()
       // Now fill the sv_eb_bndry_geom
       auto const& vfrac_arr = vfrac.array(mfi);
       auto const& bndrycent_arr = bndrycent->array(mfi);
-      AMREX_D_TERM(auto const& eb2areafrac_arr_0 = eb2areafrac[0]->array(mfi);
-                   , auto const& eb2areafrac_arr_1 = eb2areafrac[1]->array(mfi);
-                   ,
-                   auto const& eb2areafrac_arr_2 = eb2areafrac[2]->array(mfi);)
+      AMREX_D_TERM(auto const& areafrac_arr_0 = areafrac[0]->array(mfi);
+                   , auto const& areafrac_arr_1 = areafrac[1]->array(mfi);
+                   , auto const& areafrac_arr_2 = areafrac[2]->array(mfi);)
       pc_fill_sv_ebg(
         tbox, Ncut, vfrac_arr, bndrycent_arr,
-        AMREX_D_DECL(eb2areafrac_arr_0, eb2areafrac_arr_1, eb2areafrac_arr_2),
+        AMREX_D_DECL(areafrac_arr_0, areafrac_arr_1, areafrac_arr_2),
         sv_eb_bndry_geom[iLocal].data());
 
       sv_eb_bndry_grad_stencil[iLocal].resize(Ncut);
@@ -204,7 +202,7 @@ PeleC::initialize_eb2_structs()
 
       if (typ == amrex::FabType::regular || typ == amrex::FabType::covered) {
       } else if (typ == amrex::FabType::singlevalued) {
-        const auto afrac_arr = (*eb2areafrac[dir])[mfi].array();
+        const auto afrac_arr = (*areafrac[dir])[mfi].array();
         const auto facecent_arr = (*facecent[dir])[mfi].array();
 
         // This used to be an std::set for cut_faces (it ensured
@@ -651,4 +649,88 @@ PeleC::extend_signed_distance(
     }
     signDist->FillBoundary(parent->Geom(0).periodicity());
   }
+}
+
+void
+PeleC::InitialRedistribution(
+  const amrex::Real time,
+  const amrex::Vector<amrex::BCRec> bcs,
+  amrex::MultiFab& S_new)
+{
+  BL_PROFILE("PeleC::InitialRedistribution()");
+
+  // Don't redistribute if there is no EB or if the redistribution type is
+  // anything other than StateRedist
+  if (
+    (!eb_in_domain) ||
+    ((eb_in_domain) && (redistribution_type != "StateRedist"))) {
+    return;
+  }
+
+  if (verbose != 0) {
+    amrex::Print() << "Doing initial redistribution... " << std::endl;
+  }
+
+  // Initial data are set at new time step
+  amrex::MultiFab tmp(
+    grids, dmap, S_new.nComp(), numGrow(), amrex::MFInfo(), Factory());
+
+  amrex::MultiFab::Copy(tmp, S_new, 0, 0, S_new.nComp(), S_new.nGrow());
+  FillPatch(*this, tmp, numGrow(), time, State_Type, 0, S_new.nComp());
+  EB_set_covered(tmp, 0.0);
+
+  amrex::Gpu::DeviceVector<amrex::BCRec> d_bcs(bcs.size());
+  amrex::Gpu::copy(
+    amrex::Gpu::hostToDevice, bcs.begin(), bcs.end(), d_bcs.begin());
+
+  for (amrex::MFIter mfi(S_new, amrex::TilingIfNotGPU()); mfi.isValid();
+       ++mfi) {
+    const amrex::Box& bx = mfi.validbox();
+
+    auto const& fact =
+      dynamic_cast<amrex::EBFArrayBoxFactory const&>(S_new.Factory());
+
+    auto const& flags = fact.getMultiEBCellFlagFab()[mfi];
+    amrex::Array4<const amrex::EBCellFlag> const& flag_arr =
+      flags.const_array();
+
+    if (
+      (flags.getType(amrex::grow(bx, 1)) != amrex::FabType::covered) &&
+      (flags.getType(amrex::grow(bx, 1)) != amrex::FabType::regular)) {
+      amrex::Array4<const amrex::Real> AMREX_D_DECL(fcx, fcy, fcz), ccc,
+        AMREX_D_DECL(apx, apy, apz);
+
+      AMREX_D_TERM(fcx = facecent[0]->const_array(mfi);
+                   , fcy = facecent[1]->const_array(mfi);
+                   , fcz = facecent[2]->const_array(mfi););
+
+      ccc = fact.getCentroid().const_array(mfi);
+
+      AMREX_D_TERM(apx = areafrac[0]->const_array(mfi);
+                   , apy = areafrac[1]->const_array(mfi);
+                   , apz = areafrac[2]->const_array(mfi););
+
+      const auto& sarr = S_new.array(mfi);
+      const auto& tarr = tmp.array(mfi);
+      Redistribution::ApplyToInitialData(
+        bx, NVAR, sarr, tarr, flag_arr, AMREX_D_DECL(apx, apy, apz),
+        vfrac.const_array(mfi), AMREX_D_DECL(fcx, fcy, fcz), ccc,
+        d_bcs.dataPtr(), geom, redistribution_type, eb_srd_max_order);
+
+      // Make sure rho is same as sum rhoY after redistribution
+      amrex::ParallelFor(
+        bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
+          amrex::Real drhoYsum = 0.0;
+          for (int n = 0; n < NUM_SPECIES; n++) {
+            drhoYsum += sarr(i, j, k, UFS + n) - tarr(i, j, k, UFS + n);
+          }
+          sarr(i, j, k, URHO) = tarr(i, j, k, URHO) + drhoYsum;
+        });
+      amrex::ParallelFor(
+        bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
+          pc_check_initial_species(i, j, k, sarr);
+        });
+    }
+  }
+  set_body_state(S_new);
 }
